code: suc_RL_controller
This controller succeeded in creating a load planning pattern in time-SoC-space, a simplified space though. Time and SoC state have been shrunk to small portions so that they could be reached.  Random SoCs are generated during the training process.
Next step: find a way to compare it with the old result.